requirements.txt
addict
configobj
logtool
pyver
path.py
clip.py
